feat(sync): expose IBKR in the Add account modal
The IBKR (Interactive Brokers, Flex Web Service) connector shipped end-to-end
in 1.0.15, but the "Add account" modal keeps its own hardcoded SOURCES list,
which never got the IBKR entry — so IBKR was only reachable from the Sync tab.

Extract the IBKR connection UI from IbkrTab into a reusable IbkrPanel (single
source of truth), turn IbkrTab into a one-line re-export (mirroring the other
providers), and wire an IBKR entry into the Add account modal between DEGIRO
and Amundi. IbkrPanel calls onConnected on a successful connect so the modal
closes, matching the degiro/amundi/bourso pattern. Adds the addAccount.desc.ibkr
key in all four locales, a vitest mock + flow test, and a Playwright spec
proving the entry end-to-end.
