+++ frontend/src/components/shared/AddAccountModal.test.tsx
@@ -60,6 +60,12 @@ vi.mock('@/components/sync/AmundiPanel', () => ({
   ),
 }))
 
+vi.mock('@/components/sync/IbkrPanel', () => ({
+  IbkrPanel: ({ onConnected }: { onConnected?: () => void }) => (
+    <button onClick={onConnected}>ibkr-wizard</button>
+  ),
+}))
+
 vi.stubGlobal('ResizeObserver', class {
   observe() {}
   unobserve() {}
@@ -265,6 +271,18 @@ describe('AddAccountModal Amundi wizard', () => {
 
     fireEvent.click(screen.getByText('sync.amundi.title'))
     fireEvent.click(screen.getByRole('button', { name: 'amundi-wizard' }))
+
+    expect(onOpenChange).toHaveBeenCalledWith(false)
+  })
+})
+
+describe('AddAccountModal IBKR wizard', () => {
+  it('opens the connector and closes after authentication', () => {
+    const onOpenChange = vi.fn()
+    render(<AddAccountModal open onOpenChange={onOpenChange} />)
+
+    fireEvent.click(screen.getByText('sync.ibkr.title'))
+    fireEvent.click(screen.getByRole('button', { name: 'ibkr-wizard' }))
 
     expect(onOpenChange).toHaveBeenCalledWith(false)
   })
