+++ frontend/src/i18n/locales/es.json
@@ -435,6 +435,7 @@
       "bourso": "Cuentas corrientes, ahorro, PEA y cuentas de valores en BoursoBank",
       "bourseDirect": "Sincronizar tus cuentas PEA y de valores",
       "degiro": "Sincronizar tu cuenta de valores DEGIRO",
+      "ibkr": "Sincroniza tu cuenta de Interactive Brokers mediante Flex Query",
       "finary": "Importar desde Finary",
       "manual": "Crear una cuenta manualmente",
       "amundi": "PEE, PEG, PERCO, PER en Amundi",
