+++ frontend/src/i18n/locales/fr.json
@@ -435,6 +435,7 @@
       "bourso": "Comptes courants, livrets, PEA et comptes-titres chez BoursoBank",
       "bourseDirect": "Synchroniser vos PEA et CTO",
       "degiro": "Synchroniser votre compte-titres DEGIRO",
+      "ibkr": "Synchronisez votre compte Interactive Brokers via Flex Query",
       "finary": "Importer depuis Finary",
       "manual": "Créer un compte manuellement",
       "amundi": "PEE, PEG, PERCO, PER chez Amundi",
